Improve error handling in RenderSystem and update buffer size types
Enhanced error reporting for shader compilation and vertex buffer creation in RenderSystem by adding detailed debug output and exception handling. Changed vertex buffer shader byte size parameter from UINT to size_t for better type safety. Removed unused getDevice() method from RenderSystem.

+++ GDENG03 Getting Started Game Engine/VertexBuffer.h
@@ -6,7 +6,7 @@
 class VertexBuffer
 {
 public:
-	VertexBuffer(RenderSystem* system, void* list_vertices, UINT size_vertex, UINT size_list, void* shader_byte_code, UINT size_byte_shader);
+	VertexBuffer(RenderSystem* system, void* list_vertices, UINT size_vertex, UINT size_list, void* shader_byte_code, size_t size_byte_shader);
 	UINT getSizeVertexList();
 	~VertexBuffer();
 private:
